+++ server.js
@@ -7,7 +7,7 @@ const fastify_static = require('fastify-static');
 const fastify_ws = require('fastify-ws');
 
 //Server
-const version = "0.9.1";
+const version = "0.9.2";
 const settings = require('./settings.json');
 
 var Clexi = function(customSettings){